Add files via upload
The backend works in this version of the code. index.html shows list of articles, login.html will redirect to article_form.html upon successful login, then article_form.html creates an article for a logged in user and then redirects to article.detail.html. I also made a few changes to the Sass to clean up the styles and created a comics.html page that will hopefully contain a working flexslider.